Organize unit test code
