+++ tests/unit/test_cisco_commands.c
@@ -1,4 +1,4 @@
-#include "test_framework.h"
+#include "../framework/test_framework.h"
 #include "../include/cisco_cli.h"
 
 // Mock the serial_read_until function by defining it here
